# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "letmepost.dev API",
    "version": "v1",
    "description": "Open-source social media publishing API. Preflight validation, transparent errors, idempotency, stable platform versions.\n\nThe full developer documentation lives at [letmepost.dev/docs/](https://letmepost.dev/docs/). This reference is generated from the canonical Zod schemas at `packages/schemas/`.\n\n**Auth:** Bearer key in the `Authorization` header. See [Authentication](https://letmepost.dev/docs/authentication/).\n**Idempotency:** every write accepts `Idempotency-Key`. See [Idempotency](https://letmepost.dev/docs/idempotency/).\n**Errors:** stable envelope with `code`, `rule`, `platformResponse`, `remediation`, `docUrl`, `ruleUrl`. See [Errors](https://letmepost.dev/docs/errors/).",
    "contact": {
      "name": "letmepost.dev",
      "url": "https://github.com/rosekamallove/letmepost.dev"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "https://api.letmepost.dev",
      "description": "Production"
    }
  ],
...
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "lmp_live_… or lmp_test_…",
        "description": "Mint an API key in the dashboard. See https://letmepost.dev/docs/authentication/."
      }
    },
    "headers": {
      "X-RateLimit-Limit": {
        "description": "Per-route request ceiling for this endpoint. A static contract — sliding-window remaining/reset live on the IETF-draft `RateLimit-*` headers (no prefix) when enforcement is active.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 1000
        }
      },
      "X-Request-Id": {
        "description": "Per-request correlation id, also echoed inside the error envelope as `error.requestId`.",
        "schema": {
          "type": "string",
          "example": "req_01HY6X4AWBJM2K9F2PTQMRD9JQ"
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Recommended on every write. Replays within 24 hours return the original response; conflicting bodies surface as `409 idempotency_conflict`. See [Idempotency](https://letmepost.dev/docs/idempotency/).",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "example": "post-launch-2026-05-15-001"
        }
      }
    },
    "schemas": {
      "Platform": {
        "type": "string",
        "enum": [
          "bluesky",
          "facebook",
          "instagram",
          "linkedin",
          "pinterest",
          "threads",
          "twitter"
        ],
        "description": "Supported platforms in v1."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_failed",
          "preflight_failed",
          "platform_auth_failed",
          "platform_rejected",
          "platform_unavailable",
          "platform_not_enabled",
          "internal_error",
          "unauthenticated",
          "unauthorized",
          "not_found",
          "idempotency_conflict",
          "rate_limited"
        ],
        "description": "Canonical error codes. See https://letmepost.dev/docs/errors/ for per-code documentation."
      },
      "ApiError": {
        "type": "object",
        "description": "Stable error envelope. Every non-2xx response uses this shape. `docUrl` is always present; `ruleUrl` is present whenever `rule` is set.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "docUrl",
              "requestId"
            ],
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string",
                "description": "Human-readable summary. Stable enough to log, not stable enough to switch on — switch on `code` and `rule`."
              },
              "rule": {
                "type": "string",
                "description": "The specific preflight rule or validator that failed (e.g. `bluesky.text.max_graphemes`, `targets.options.platform_mismatch`).",
                "example": "bluesky.text.max_graphemes"
              },
              "platform": {
                "$ref": "#/components/schemas/Platform"
              },
              "platformVersion": {
                "type": "string",
                "description": "Pinned upstream API version the call targeted (e.g. `graph-v23.0` for Facebook)."
              },
              "platformResponse": {
                "description": "Raw upstream platform response, when available."
              },
              "remediation": {
                "type": "string",
                "description": "Actionable next step for the caller — short, imperative."
              },
              "docUrl": {
                "type": "string",
                "format": "uri",
                "description": "Link to the docs page for this `code`.",
                "example": "https://docs.letmepost.dev/errors/preflight_failed"
              },
              "ruleUrl": {
                "type": "string",
                "format": "uri",
                "description": "Link to the docs page for this preflight `rule`. Present whenever `rule` is set.",
                "example": "https://docs.letmepost.dev/preflight/bluesky-text-max_graphemes"
              },
              "requestId": {
                "type": "string",
                "description": "Per-request correlation id, also echoed in the `X-Request-Id` response header.",
                "example": "req_01HY6X4AWBJM2K9F2PTQMRD9JQ"
              },
              "traceId": {
                "type": "string",
                "description": "OpenTelemetry trace id, when tracing is active."
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "preflight_failed",
            "message": "Bluesky posts are capped at 300 graphemes; this body is 312.",
            "rule": "bluesky.text.max_graphemes",
            "platform": "bluesky",
            "remediation": "Trim 12 graphemes, or split into a thread.",
            "docUrl": "https://docs.letmepost.dev/errors/preflight_failed",
            "ruleUrl": "https://docs.letmepost.dev/preflight/bluesky-text-max_graphemes",
            "requestId": "req_01HY6X4AWBJM2K9F2PTQMRD9JQ"
          }
        }
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "ci-publish-key"
          },
          "prefix": {
            "type": "string",
            "enum": [
              "lmp_live_",
              "lmp_test_"
            ],
            "default": "lmp_live_"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "profileId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional profile scope. When set, the key only authenticates requests targeting that profile; cross-profile reads/writes 404."
          }
        }
      },
      "CreateApiKeyResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "last4",
          "scopes",
          "key",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "last4": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "profileId": {
            "type": [
              "string",
              "null"
            ]
          },
          "key": {
            "type": "string",
            "description": "Plaintext key — returned **once** at creation and never again. Store it securely.",
            "example": "lmp_live_AbC123…"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "last4",
          "scopes",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "last4": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "profileId": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "ApiKeyListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            }
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "`validation_failed` — request body or query parameters didn't match the schema. Rules surfaced include `body.json`, `targets.required`, `targets.max`, `mode_conflict`, `targets.options.platform_mismatch`, `targets.account.platform_mismatch`, `target.account.not_connected`, `target.account.ambiguous`, `scheduledAt.future`, `scheduledAt.text_only`.",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "Unauthenticated": {
        "description": "`unauthenticated` — missing or invalid Bearer key.",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found` — the resource does not exist in the caller's organization (or, for profile-scoped keys, in the caller's profile).",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "IdempotencyConflict": {
        "description": "`idempotency_conflict` — the same `Idempotency-Key` was used with a different request body. Use a new key, or retry with the original body.",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "`rate_limited` — too many requests. Inspect `X-RateLimit-Limit` (static ceiling) and the IETF-draft `RateLimit-*` headers (sliding-window remaining/reset) before retrying.",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "`internal_error` — something on our end failed unexpectedly. Quote the `requestId` when reporting.",
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/X-Request-Id"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Mint an API key",
        "description": "Dashboard session only. The plaintext `key` is returned **once** in this response and never again — store it securely. Optional `profileId` pins the key to a single profile; cross-profile reads/writes 404 under that key.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              },
              "example": {
                "name": "ci-publish-key",
                "prefix": "lmp_live_",
                "scopes": []
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Key minted — plaintext shown once.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "description": "Dashboard session only. Plaintext is never returned again — only `last4` and metadata.",
        "security": [],
        "responses": {
          "200": {
            "description": "Active and revoked keys for the active org.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke an API key",
        "description": "Soft-revoke. The key is invalidated immediately; the row stays for audit. Already-revoked keys 404.",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "revokedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "revokedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
...
}